+++ src/toKana.js
@@ -1,6 +1,4 @@
-import {
-  DEFAULT_OPTIONS,
-} from './constants';
+import { DEFAULT_OPTIONS } from './constants';
 import {
   getRomajiToKanaTree,
   IME_MODE_MAP,
@@ -31,24 +29,25 @@ import hiraganaToKatakana from './utils/hiraganaToKatakana';
  */
 export function toKana(input = '', options = {}) {
   // just throw away the substring index information and just concatenate all the kana
-  return splitIntoKana(input, options).map((kanaToken) => {
-    const [start, , kana] = kanaToken;
-    if (kana === null) {
-      // haven't converted the end of the string, since we are in IME mode
-      return input.slice(start);
-    }
-    // make katakana, if the first letter of the syllable is upper case
-    const makeKatakana = !options.ignoreCase && isCharUpperCase(input.charAt(start));
-    return makeKatakana? hiraganaToKatakana(kana): kana;
-  }).join('');
+  return splitIntoKana(input, options)
+    .map((kanaToken) => {
+      const [start, , kana] = kanaToken;
+      if (kana === null) {
+        // haven't converted the end of the string, since we are in IME mode
+        return input.slice(start);
+      }
+      // make katakana, if the first letter of the syllable is upper case
+      return isCharUpperCase(input.charAt(start)) ? hiraganaToKatakana(kana) : kana;
+    })
+    .join('');
 }
 
 export function splitIntoKana(input = '', options = {}) {
   const config = Object.assign({}, DEFAULT_OPTIONS, options);
 
   let map = getRomajiToKanaTree(config);
-  map = config.IMEMode? IME_MODE_MAP(map): map;
-  map = config.useObsoleteKana? USE_OBSOLETE_KANA_MAP(map): map;
+  map = config.IMEMode ? IME_MODE_MAP(map) : map;
+  map = config.useObsoleteKana ? USE_OBSOLETE_KANA_MAP(map) : map;
   map = config.customKanaMapping(map);
   return applyMapping(input.toLowerCase(), map, !config.IMEMode);
 }