+++ CHANGELOG.md
@@ -22,9 +22,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Add any unpublished changes here as they are made, for easy reference come release time.
 -->
+## [2.3.4] - 2017-12-16
+### Fixed
+- *Uppercase* input with toHiragana() including a double consonant was incorrectly producing a katakana ッ instead of っ
+
+## [2.3.3] - 2017-12-12
+### Fixed
+- Add https to unpkg link
+
+## [2.3.2] - 2017-12-12
+### Added
+- Specify minified bundle in package.json for easy use with unpkg
+
 ## [2.3.1] - 2017-10-17
 ### Changed
-- Set cursor in IME mode to the correct position after conversion
+- Set cursor in IME mode to the current position after conversion (rather than the end of all input)
 
 ## [2.3.0] - 2017-08-28
 ### Changed
@@ -104,6 +116,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `isRomaji()` allows hepburn romanisation long vowels. (IE. Tōkyō)
 
 
+[2.3.4]: https://github.com/WaniKani/WanaKana/compare/2.3.3...2.3.4
+[2.3.3]: https://github.com/WaniKani/WanaKana/compare/2.3.2...2.3.3
+[2.3.2]: https://github.com/WaniKani/WanaKana/compare/2.3.1...2.3.2
+[2.3.1]: https://github.com/WaniKani/WanaKana/compare/2.3.0...2.3.1
 [2.3.0]: https://github.com/WaniKani/WanaKana/compare/2.2.4...2.3.0
 [2.2.4]: https://github.com/WaniKani/WanaKana/compare/2.2.3...2.2.4
 [2.2.3]: https://github.com/WaniKani/WanaKana/compare/2.2.2...2.2.3